✨ NEW: Basic에 값 넣기 & Basic 값 전부 가져오기

+++ src/main/java/com/bside/whatMeal/domain/entity/FoodIngredient.java
@@ -14,7 +14,8 @@ public class FoodIngredient {
     //음식 이름과 주재료 선택 관계 테이블
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+
+    private int id;
 
     @ManyToOne
     @JoinColumn(name = "food_id")
